requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.8.0
werkzeug>=2.3.3
firebase-admin>=6.0.0
google-cloud-firestore==2.13.1
google-generativeai==0.3.2
plaid-python==17.0.0
pydantic>=1.10.0
